+- update patches.suse/netfilter-06-policy-match +  ->match passes int, not u16
suse-commit: 80d347baff61eb35ec4e11384206fc0243a766af

+++ net/ipv6/netfilter/ip6t_policy.c
@@ -121,7 +121,7 @@ static int match(const struct sk_buff *skb,
                  const void *matchinfo,
 		 int offset,
 		 const void *hdr,
-		 u_int16_t datalen,
+		 int datalen,
 		 int *hotdrop)
 {
 	const struct ip6t_policy_info *info = matchinfo;
